GH-1003 Move TokenResponse to auth package
