+++ region-connectors/region-connector-fr-enedis/src/main/java/energy/eddie/regionconnector/fr/enedis/dto/auth/TokenResponse.java
@@ -1,4 +1,4 @@
-package energy.eddie.regionconnector.fr.enedis.dto;
+package energy.eddie.regionconnector.fr.enedis.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
